Check if running in NVIDIA graphics mode
Just checking for the proprietary drivers is no longer sufficient, as
they will be loaded in hybrid graphics mode but not be used for
rendering the desktop.

Rely on the `/etc/prime-discrete` file written by system76-power and
gpu-manager to check which rendering mode is being used, falling back to
checking the modules if it doesn't exist.

Signed-off-by: Tim Crawford <tcrawford@system76.com>